updata jdbc to hibernate

+++ src/sample/Model/Item_fortable.java
@@ -1,5 +1,6 @@
 package sample.Model;
 
+import javafx.beans.property.SimpleFloatProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
 import org.springframework.stereotype.Component;
@@ -9,16 +10,16 @@ public class Item_fortable {
     SimpleIntegerProperty itemid;
     SimpleStringProperty item;
     SimpleIntegerProperty quality;
-    SimpleIntegerProperty price;
+    SimpleFloatProperty price;
 
     public Item_fortable() {
     }
 
-    public Item_fortable(int itemid, String item, int quality, int price) {
+    public Item_fortable(int itemid, String item, int quality, float price) {
         this.itemid =new SimpleIntegerProperty(itemid);
         this.item = new SimpleStringProperty(item);
         this.quality =new SimpleIntegerProperty(quality);
-        this.price = new SimpleIntegerProperty(price);
+        this.price = new SimpleFloatProperty(price);
     }
 
     public int getItemid() {
